Move getDataPlaneScope and parseDataPlaneName into data-plane utils
